requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
attrs==22.1.0
    # via
    #   fiona
    #   jsonschema
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via django-geostore
certifi==2022.9.24
    # via
    #   fiona
    #   requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   cligj
    #   fiona
    #   mercantile
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via
    #   celery
    #   fiona
click-repl==0.2.0
    # via celery
cligj==0.7.2
    # via fiona
deepmerge==1.0.1
    # via django-geostore
django==4.1.2
    # via
    #   -r requirements.in
    #   django-geostore
    #   django-token-tools
    #   djangorestframework
django-geostore==0.6.8
    # via -r requirements.in
django-token-tools==0.1.2
    # via django-geostore
djangorestframework==3.14.0
    # via
    #   django-geostore
    #   djangorestframework-gis
djangorestframework-gis==1.0
    # via django-geostore
fiona==1.8.22
    # via django-geostore
gpxpy==1.5.0
    # via django-geostore
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
jsonschema==4.16.0
    # via django-geostore
kombu==5.2.4
    # via celery
mercantile==1.2.1
    # via django-geostore
munch==2.5.0
    # via fiona
prompt-toolkit==3.0.31
    # via click-repl
psycopg2==2.8.6
    # via django-geostore
pyrsistent==0.18.1
    # via jsonschema
pytz==2022.5
    # via
    #   celery
    #   djangorestframework
requests==2.28.1
    # via django-geostore
simplekml==1.3.6
    # via django-geostore
six==1.16.0
    # via
    #   click-repl
    #   fiona
    #   munch
sqlparse==0.4.3
    # via django
urllib3==1.26.12
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
